.env remove
