+++ images/harvest/experts-client/package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "@comunica/bindings-factory": "^2.7.0",
     "@comunica/query-sparql": "^2.7.1",
+    "@google-cloud/secret-manager": "^4.2.2",
     "@rdfjs/fetch-lite": "3.2.1",
     "classic-level": "^1.2.0",
     "commander": "^10.0.0",
